+++ src/pages/admin/Users/Form_SubmitUser/form_submit_user.module.scss
@@ -1,0 +1,21 @@
+.heading {
+    color: #C67D39;
+}
+
+.content {
+
+    form {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        width: 100%;
+
+        .formGroup {
+            margin: 0;
+            width: 100%;
+            min-width: 100vb;
+
+            
+        }
+    }
+}
